Now resolving test & prod deps + running tests!
The test_db container remains running and doesn't terminate, but to fix
it, the Docker package needs to handle a TCP DOCKER_HOST.  I think that
this will involve setting up a local Docker package for experimentation:
https://docs.dagger.io/1010/dev-cue-package/

This helped me pick on a few Dagger structuring ideas:
https://www.youtube.com/watch?v=QvyB3m9Fti0
Thanks @talentedmrjones!

The OTEL Jaeger integration is nice, there is a gotcha in the docs which
I'm adding it to my TODO list. For now, the focus is on getting the
CircleCI pipeline replicated in Dagger. Next step: run test_db with
`docker.#Run`. Follow-up: introduce `docker#Remove`.

Current commmands:
    make dagger-ci-test
    make dagger-ci-prod

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>