First attempt and building prod image
We needed to remove the docker dir from the dockerignore so that it
would be available for the docker.#Build step.

We had to make the inputs part of the dagger-ci so that they would get
correctly re-defined when running this task. Inputs change post
environment declaration, so this feels right.

For some reason, the source in our docker.#Build is not copied
correctly, because prod deps are pulled and compiled again, which should
not happen if prod_assets were copied correctly. Also, the assets should
be compiled and digested, which apparently is not the case.

There is a slight problem with copying from prod_assets, as this also
includes the dev_deps (required for assets to be digested). Ideally, we
would combine prod_deps & /app/assets + /app/priv from prod_assets, but
using multiple sources is not currently supported.

Thanks @jlongtine & @grouville for the extra pair of eyes & hands!

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>